adding test for write old paths
